split interface to new file
